add commit function

+++ src/main/java/life/gao/community/dto/QuestionDTO.java
@@ -5,13 +5,13 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private long id;
     private String title;
     private String description;
     private String tag;
     private  long  gmtCreate;
     private  long gmtModified;
-    private Integer creator;
+    private long creator;
     private Integer  viewCount;
     private Integer  commentCount;
     private Integer  likeCount;
